Add files via upload
These are the files required for the execution of this application. The seamless functioning of this app relies on the inclusion of specific files. These essential components ensure the proper execution and performance of the application. It is imperative to have the following files in place for the app to run smoothly and efficiently.